Moved password_hash to dedicated file
* Passwords can be set for units as well as ground control
  roles therefore it's more suitable to have password_hash
  implemented as a utility.
